upsert-requirements.txt
langchain
transformers
pinecone-client
openai
tiktoken
python-dotenv
pandas
numpy